refactor(headless): split Tabs into per-file parts for tree-shaking
