+++ packages/headless/src/primitives/tabs/parts.ts
@@ -1,0 +1,6 @@
+export { type TabsProps, TabsRoot as Root } from './tabs-root';
+export { type TabsListProps, TabsList as List } from './tabs-list';
+export { type TabsTabProps, TabsTab as Tab } from './tabs-tab';
+export { type TabsTriggerProps, TabsTrigger as Trigger } from './tabs-trigger';
+export { type TabsPanelProps, TabsPanel as Panel } from './tabs-panel';
+export { type TabsIndicatorProps, TabsIndicator as Indicator } from './tabs-indicator';
